Fix the parent/child relationship in the docs
In the Docker Linking Docs, the parent child relationship
was backwards. The Web container should be able to access
the DB, not other way around. Furthermore, the output of
'docker ps' was wrong (it showed that the DB could access
the Web). This fixes both typos.

Docker-DCO-1.1-Signed-off-by: Ian Bull <irbull@gmail.com> (github: irbull)