image spec: Correct set of acceptable characters in tags
The image spec did not match the regexp that validates tags. It
neglected to mention that period and dash characters are allowed in
tags, as long as they are not the first character. It also did not
mention the length limit for tags.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit 84e5449e05f0bfa6cdeb0a8e51a7f46a572d2d48)
Signed-off-by: Tibor Vass <tibor@docker.com>