Add selinux policy per distro
This adds the ability to have different profiles for individual distros
and versions of the distro because they all ship with and depend on
different versions of policy packages.

The `selinux` dir contains the unmodified policy that is being used
today.  The `selinux-fedora` dir contains the new policy for fedora 24
with the changes for it to compile and work on the system.

The fedora policy is from commit
https://github.com/projectatomic/docker-selinux/commit/4a6ce94da5e34868642ebcc7a67da937d4c73283

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>
(cherry picked from commit be3c34407dd3ef77306811dbd9ae8788b1dabc9b)
Signed-off-by: Tibor Vass <tibor@docker.com>