Improve example for env vars in API docs
This was added in 39496ded2bb392c20411f085e8898d43d96d2477,
but can be a bit confusing due to the "outer" brackets
being for JSON, and the "inner" brackets to indicate
optional values.

During review of that change, this alternative
format was suggested.

Updating the example to use the format that was discussed
during review :)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 95d84dd2f2b6021c69910912708db9abdc60cd66)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>