Don't resolve or pull images referenced by ID
If a swarm service is created using an image ID, it's useless to try to
pull this reference or resolve it to a manifest digest. Avoid doing this
when a fully qualified image ID is given.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit 910e148f9a1b1e02f98a2029952b9ded80e39001)
Signed-off-by: Victor Vieux <victorvieux@gmail.com>