Fix out-of-band vol delete+create for same driver
Fix issue where out-of-band deletions and then a `docker volume create`
on the same driver caused volume to not be re-created in the driver but
return as created since it was stored in the cache.

Previous fix only worked if the driver names did not match.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit 2caa94b93d48e1b2863cbf87677a21ef49e1fd2f)
Signed-off-by: Victor Vieux <victorvieux@gmail.com>