Fix restartpolicy max-retry validation
the restart policy validation was moved from
the client to the daemon in b13bc1d57daf91ee57557a9171ea49e158bbee74

As part of that change, retry-counts < 1
were marked as "invalid".

However, the default is 0 (unlimited), causing

    docker run -d --restart=on-failure nginx

To fail.

This changes the validation to only invalidate
retry-counts < 0.

A test was added, and other tests renamed
to allow running just these tests :)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>